requirements.txt
asgiref==3.4.1
certifi==2022.9.14
charset-normalizer==2.0.12
Django==3.2.15
django-debug-toolbar==3.2.4
django-filter==21.1
django-utils-six==2.0
djangorestframework==3.13.1
idna==3.4
importlib-metadata==4.8.3
Pillow==8.4.0
psycopg2-binary==2.9.3
pytz==2022.2.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.12
whitenoise==5.3.0
zipp==3.6.0
